test(scheduling): extract alarm scheduling contract
